🧹 Extract security-agent types, constants, and utilities
